Refactor authentication logic to use features/authSlice; add Protected component for route protection
