+++ myApp/src/components/Signup.jsx
@@ -1,7 +1,7 @@
 import React, {useState} from 'react';
 import authService from '../services/auth.js';
 import {Link ,useNavigate} from 'react-router-dom';
-import {login} from '../store/authSlice';
+import {login} from '../features/authSlice.js';
 import {Button, Input, Logo} from './index.js';
 import {useDispatch} from 'react-redux';
 import {useForm} from 'react-hook-form';
