+++ myApp/src/features/authSlice.js
@@ -9,16 +9,16 @@ const authSlice = createSlice({
     name: "auth",
     initialState,
     reducers:{
-        logIn:(state, action) =>{
+        login:(state, action) =>{
             state.status = true,
             state.userData = action.payload.userData
         },
-        logOut:(state, action) =>{
+        logout:(state, action) =>{
             state.status = false;
             state.userData = null
         }
     }
 });
 
-export const {logIn, logOut} = authSlice.actions;
+export const {login, logout} = authSlice.actions;
 export default authSlice.reducers;
